Also refactor aws-s3-multipart
